Allocate a string that's just large enough for the Content-Length value
Previously, we were assuming that we would never encounter a response
body with a size (in bytes) that would require more than 10 characters
(including the terminating NULL) when turned into a string.

This commit uses a helper function sizetoa(), analogous to the non-
standard asprintf(3), which only allocates exactly the right amount of
memory to format a size_t as a string.